+++ changes.md
@@ -2,3 +2,4 @@
 - The company connect button no longer fetches the company details when the details are already available from it's parent
 
 - Companies in the company page are now server side sorted and paginated
+- Ordering/sorting been changed from sort_by to ordering in the api call to get companies